nu är lite av gfx portat till ogl
git-svn-id: http://muon.googlecode.com/svn/trunk@70 99aa2080-d4e9-11dd-8179-53c65a4d6d95

+++ player.h
@@ -3,7 +3,7 @@
 
 typedef struct
 {
-	Uint8 r, g, b;
+	float r, g, b;
 	int blost;
 	int brazed;
 	int bbuilt;
